changed match results page back to noramal
